test(core): colocate SVG export coverage
